requirements.txt
Django>=5.1
djangorestframework>=3.15.2
django-cors-headers>=4.9.0
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1
ruff>=0.3.0
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.2.0
django-private-chat2<=1.0.2
